Serial and Parallel Programming Using Python
This project made in python is to calculate how fast it takes to encrypt and de-crypt the password from the input file(where you put all the words to encryption file) for it to be encrypted. Once the word in encrypted, it will be converted into hashes and it will be stored in the output file.